+++ src/components/CredentialsSharingTool.vue
@@ -101,7 +101,7 @@ export default class CredentialsSharingTool extends Vue {
     this.recipientsMessageSubjects = [];
     if (this.recipientsEmails) {
       this.recipientsEmails.forEach(emailRole => {
-        this.projectManagerBackendService.isModuleActionActive(Module.PROJECT_RESULTS_MODULE, Action.FETCH_EMAIL_MESSAGE_AND_SUBJECT_ACTION).then(condition => {
+        this.projectManagerBackendService.isModuleActionActive(Module.PROJECT_RESULTS_MODULE, Action.FETCH_EMAIL_MESSAGE_AND_SUBJECT_ACTION, this.context).then(condition => {
           if (condition) {
             this.projectManagerBackendService.fetchData(Module.PROJECT_RESULTS_MODULE, Action.FETCH_EMAIL_MESSAGE_AND_SUBJECT_ACTION, this.context,
                 new Map([['email', emailRole.email], ['project-role', emailRole.projectRole], ['email-template-type', this.fetchEmailTemplateType()]])).then(messageSubject => {