+++ src/components/NavBar.vue
@@ -78,7 +78,7 @@ export default defineComponent({
 
     async fetchProjectRoles() {
       try {
-        this.projectManagerBackendService.isModuleActionActive(Module.USER_MODULE, Action.IS_PROJECT_MANAGER_ADMIN_ACTION).then(condition => {
+        this.projectManagerBackendService.isModuleActionActive(Module.USER_MODULE, Action.IS_PROJECT_MANAGER_ADMIN_ACTION, this.context).then(condition => {
           this.isProjectManagerAdmin = false;
           if (condition) {
             this.projectManagerBackendService.fetchData(Module.USER_MODULE, Action.IS_PROJECT_MANAGER_ADMIN_ACTION, this.context, new Map()).then(isProjectManagerAdmin => {